backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# AI/ML Framework
langgraph==0.0.66
groq==0.4.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
aiofiles==23.2.0

# Embeddings and Vector Search
sentence-transformers==2.2.2
numpy==1.25.2
scikit-learn==1.3.2

# HTTP client
httpx==0.25.2

# Email validation
email-validator==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# File handling
Pillow==10.1.0

# Development tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1